Move current implementation into a library package
- I'm calling this the "runner" style. I have no idea what actual already-named pattern I'm following. There's probably a better name.
